Refactored directory names
Refactored directory names to make more sense under the proposed SSH Cloud architecture.
