--- Tables/src/main/java/com/example/repositories/BasketItemRepository.java
@@ -1,48 +1,0 @@
-package com.example.repositories;
-
-import com.example.entities.BasketItem;
-import com.example.entities.embeddables.BasketItemId;
-import org.hibernate.Session;
-import org.hibernate.SessionFactory;
-import org.hibernate.Transaction;
-
-import java.util.List;
-
-public class BasketItemRepository {
-    private final SessionFactory sessionFactory;
-
-    public BasketItemRepository(SessionFactory sessionFactory) {
-        this.sessionFactory = sessionFactory;
-    }
-
-    public void create(BasketItem basketItem) {
-        try (Session session = sessionFactory.openSession()) {
-            Transaction transaction = session.beginTransaction();
-            session.save(basketItem);
-            transaction.commit();
-        }
-    }
-
-    public List<BasketItem> getAll() {
-        try (Session session = sessionFactory.openSession()) {
-            return session.createQuery("from BasketItem", BasketItem.class).list();
-        }
-    }
-
-    public BasketItem getById(BasketItemId id) {
-        try (Session session = sessionFactory.openSession()) {
-            return session.get(BasketItem.class, id);
-        }
-    }
-
-    public void deleteById(BasketItemId id) {
-        try (Session session = sessionFactory.openSession()) {
-            Transaction transaction = session.beginTransaction();
-            BasketItem basketItem = session.get(BasketItem.class, id);
-            if (basketItem != null) {
-                session.delete(basketItem);
-            }
-            transaction.commit();
-        }
-    }
-}
