+++ Tables/src/main/java/ssh/entities/Store.java
@@ -1,0 +1,27 @@
+package ssh.entities;
+
+import javax.persistence.*;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Getter;
+import lombok.Setter;
+@Getter
+@Setter
+@Entity
+@Table(name = "Store")
+public class Store {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @JsonProperty("store_id")
+    private int storeId;
+
+    @Column(name = "store_name", nullable = false)
+    @JsonProperty("store_name")
+    private String storeName;
+
+    @Column(name = "store_logo")
+    @JsonProperty("store_logo")
+    private String storeLogo;
+
+    // Getters and Setters
+}
